Refactor resnet to imagenet + clean up
