+++ workloads/imagenet/imagenet_pytorch/resnet.py
