+++ workloads/imagenet/imagenet_pytorch/submission.py
@@ -4,8 +4,9 @@ import spec
 
 from workloads.mnist.mnist_pytorch.submission import update_params, data_selection
 
+
 def get_batch_size(workload_name):
-  batch_sizes = {'resnet_pytorch': 128}
+  batch_sizes = {'imagenet_pytorch': 128}
   return batch_sizes[workload_name]
 
 
@@ -20,9 +21,9 @@ def init_optimizer_state(
   del workload
 
   optimizer_state = {
-    'optimizer': torch.optim.SGD(model_params.parameters(),
-                                  lr=hyperparameters.learning_rate,
-                                  momentum=hyperparameters.momentum,
-                                  weight_decay=hyperparameters.weight_decay)
+      'optimizer': torch.optim.SGD(model_params.parameters(),
+                                   lr=hyperparameters.learning_rate,
+                                   momentum=hyperparameters.momentum,
+                                   weight_decay=hyperparameters.weight_decay)
   }
   return optimizer_state
