+++ workloads/imagenet/imagenet_pytorch/workload.py
@@ -5,24 +5,117 @@ import os
 from collections import OrderedDict
 from typing import Tuple
 
-from torchvision.datasets.folder import ImageFolder
-
-import spec
 import torch
 import torch.nn.functional as F
 from torchvision import transforms
 from torchvision.datasets import ImageNet
-from workloads.resnet.resnet_pytorch.resnet import resnet50
-from workloads.resnet.workload import Resnet
-from .utils import fast_collate
+from torchvision.datasets.folder import ImageFolder
+
+import spec
+import random_utils as prng
+from workloads.imagenet.imagenet_pytorch.resnet import resnet50
+from workloads.imagenet.imagenet_pytorch.utils import fast_collate
+
 
 DEVICE='cuda' if torch.cuda.is_available() else 'cpu'
 
 
-class ResnetWorkload(Resnet):
+class ImagenetWorkload(spec.Workload):
 
   def __init__(self):
     self._eval_ds = None
+
+  def has_reached_goal(self, eval_result: float) -> bool:
+    return eval_result['accuracy'] > self.target_value
+
+  @property
+  def target_value(self):
+    return 0.76
+
+  @property
+  def loss_type(self):
+    return spec.LossType.SOFTMAX_CROSS_ENTROPY
+
+  @property
+  def train_mean(self):
+    return (0.485, 0.456, 0.406)
+
+  @property
+  def train_stddev(self):
+    return (0.229, 0.224, 0.225)
+
+  """ data augmentation settings """
+  @property
+  def scale_ratio_range(self):
+    return (0.08, 1.0)
+
+  @property
+  def aspect_ratio_range(self):
+    return (0.75, 4.0 / 3.0)
+
+  @property
+  def center_crop_size(self):
+    return 224
+
+  @property
+  def resize_size(self):
+    return 256
+
+  @property
+  def num_train_examples(self):
+    return 1271167
+
+  @property
+  def num_eval_examples(self):
+    return 100000
+
+  @property
+  def max_allowed_runtime_sec(self):
+    return 111600  # 31 hours
+
+  @property
+  def eval_period_time_sec(self):
+    return 6000  # 100 mins
+
+  def _eval_metric(self, logits, labels):
+    """Return the mean accuracy and loss as a dict."""
+    raise NotImplementedError
+
+  def eval_model(
+      self,
+      params: spec.ParameterContainer,
+      model_state: spec.ModelAuxiliaryState,
+      rng: spec.RandomState,
+      data_dir: str):
+    """Run a full evaluation of the model."""
+    data_rng, model_rng = prng.split(rng, 2)
+    eval_batch_size = 128
+    if self._eval_ds is None:
+      self._eval_ds = self.build_input_queue(
+          data_rng, 'test', data_dir, batch_size=eval_batch_size)
+
+    total_metrics = {
+        'accuracy': 0.,
+        'loss': 0.,
+    }
+    num_batches = 0
+    for (images, labels) in self._eval_ds:
+      images, labels = self.preprocess_for_eval(images, labels, None, None)
+      logits, _ = self.model_fn(
+          params,
+          images,
+          model_state,
+          spec.ForwardPassMode.EVAL,
+          model_rng,
+          update_batch_norm=False)
+      # TODO(znado): add additional eval metrics?
+      batch_metrics = self._eval_metric(logits, labels)
+      total_metrics = {
+          k: v + batch_metrics[k] for k, v in total_metrics.items()
+      }
+      num_batches += 1
+
+    return {k: float(v / num_batches) for k, v in total_metrics.items()}
 
   def _build_dataset(self,
       data_rng: spec.RandomState,
@@ -67,7 +160,6 @@ class ResnetWorkload(Resnet):
 
     return dataloader
 
-
   def build_input_queue(
       self,
       data_rng: spec.RandomState,
@@ -84,7 +176,7 @@ class ResnetWorkload(Resnet):
     raise NotImplementedError
 
   def model_params_types(self):
-    raise NotImplementedError
+    pass
 
   # Return whether or not a key in spec.ParameterTree is the output layer
   # parameters.
@@ -120,7 +212,7 @@ class ResnetWorkload(Resnet):
   def model_fn(
       self,
       params: spec.ParameterContainer,
-      augmented_and_preprocessed_input_batch: spec.Tensor,
+      input_batch: spec.Tensor,
       model_state: spec.ModelAuxiliaryState,
       mode: spec.ForwardPassMode,
       rng: spec.RandomState,
@@ -140,7 +232,7 @@ class ResnetWorkload(Resnet):
     }
 
     with contexts[mode]():
-      logits_batch = model(augmented_and_preprocessed_input_batch)
+      logits_batch = model(input_batch)
 
     return logits_batch, None
